+++ .eslintrc.js
@@ -8,8 +8,5 @@ module.exports = {
   parserOptions: {
     ecmaVersion: 12,
   },
-  rules: {
-    // 'comma-dangle': 'off',
-    // 'space-before-function-paren': 'off',
-  },
+  rules: {},
 }
